+++ actuator/src/main/java/org/tron/core/actuator/ShieldedTransferActuator.java
@@ -23,6 +23,7 @@ import org.tron.common.zksnark.LibrustzcashParam.FinalCheckParams;
 import org.tron.common.zksnark.MerkleContainer;
 import org.tron.core.capsule.AccountCapsule;
 import org.tron.core.capsule.BytesCapsule;
+import org.tron.core.capsule.TransactionCapsule;
 import org.tron.core.capsule.TransactionResultCapsule;
 import org.tron.core.exception.BalanceInsufficientException;
 import org.tron.core.exception.ContractExeException;
@@ -282,7 +283,7 @@ public class ShieldedTransferActuator extends AbstractActuator {
     }
 
     long shieldedTransactionFee = calcFee();
-    byte[] signHash = TransactionUtil.getShieldTransactionHashIgnoreTypeException(tx.getInstance());
+    byte[] signHash = TransactionCapsule.getShieldTransactionHashIgnoreTypeException(tx.getInstance());
 
     if (CollectionUtils.isNotEmpty(spendDescriptions)
         || CollectionUtils.isNotEmpty(receiveDescriptions)) {
